Added Expectations, Methods and more Code to create a final Table and Graph
The original Graph did not work because the data set was to large so I added the code given in some of the worksheets to work with it. I tried to make a graph outside of class once and that the only reason why I knew to add that line of code.