Created Project Report and added code to Methods Section
Hey everyone,
I've created the project report so feel free to edit it!
This is kinda a rough draft of the code for the methdos section. I know that the code needs comments and that the value for the knn is pretty weird, so it needs some more work.
I mostly just wanted to commit this so you guys can start working on it.
Thanks!